1b was missing first sample
Array job in 1b was not running the first sample due to the format of the script. Previously the master file had no header so when SGE_TASK_ID=2 (the first iteration) the script on the second line would be run, missing the script on the first line. 
Now the master script has an innocuous header line so the second line corresponds to the first script to be run. I also had to adjust the number of scripts to be run (njobs1b variable) so that all the jobs in a script would run.